requirements.txt
django==3.2.3
django-filter
ipython
selenium
django-crispy-forms
<<<<<<< HEAD
django[argon2]
django[bcrypt]
=======
>>>>>>> 629dcdea52361b94bcb1f3203b893bd099d51585
